+++ service/controller/resource/monitoring/remotewriteapiendpointconfigsecret/resource.go
@@ -138,9 +138,9 @@ func toSecret(ctx context.Context, v interface{}, config Config) (*corev1.Secret
 
 func defaultQueueConfig() promv1.QueueConfig {
 	return promv1.QueueConfig{
-		Capacity:          10000,
-		MaxSamplesPerSend: 1000,
-		MaxShards:         50,
+		Capacity:          30000,
+		MaxSamplesPerSend: 10000,
+		MaxShards:         10,
 	}
 }