+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Generate prometheus remote write agent secret and config.
 - Configure prometheus remote write agent ingress.
 - Add Slack channel for Team Shield.
+- Add scraping of Cilium on Management Clusters.
 
 ## [4.6.2] - 2022-09-13