+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add scraping of Cilium on Management Clusters.
 
+### Fixed
+
+- Fix remote write endpoint ingress buffer size to avoid the use of temporary buffer file in the ingress controller.
+
+
 ## [4.6.4] - 2022-10-17
 
 ### Changed