+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add scraping of Cilium on Management Clusters.
 
+### Changed
+
+- Configure working queue config for the remote write receiver (reducing max number of shards but increasing sample capacity).
+
 ## [4.6.4] - 2022-10-17
 
 ### Changed